[Rename 🏷️] 이름 변경 오류 수정
