+++ src/components/PostDetail/components/PostDetailInfo/PostDetailInfo.tsx
@@ -1,6 +1,6 @@
 import * as S from "./PostDetailInfo.Styles"
 import { Post } from "@/types"
-import PostDetailUser from "./components/PostDetailUser/PostDetailUser.Styles"
+import PostDetailUser from "./components/PostDetailUser/PostDetailUser"
 import PostDetailStatus from "./components/PostDetailStatus/PostDetailStatus"
 import PostDetailComment from "./components/PostDetailComment/PostDetailComment"
