+++ src/components/PostDetail/components/PostDetailInfo/components/PostDetailUser/PostDetailUser.Styles.ts
